leds: Add led_set_brightness_sync to the public LED subsystem API
led_set_brightness_sync function was visible only internally to the
LED subsystem. It is now being made publicly available since it has
become apparent that this is a caller who should decide whether
brightness is to be set in a synchronous or an asynchronous way.

The function is modified to use brightness_set_nonblocking or
brightness_set op if brightness_set_sync is not implemented.
Eventually all LED subsystem drivers will be modfified to set brightness
only in a synchronous way with use of brightness_set_nonblocking or
brightness_set op and brightness_set_sync op will be removed. LED core
will take care of delegating brightness_set op to the work queue task in
specific contexts.

Signed-off-by: Jacek Anaszewski <j.anaszewski@samsung.com>
Cc: Bryan Wu <cooloney@gmail.com>
Cc: Andrew Lunn <andrew@lunn.ch>
Cc: Sakari Ailus <sakari.ailus@linux.intel.com>
Cc: Pavel Machek <pavel@ucw.cz>
Cc: Stas Sergeev <stsp@users.sourceforge.net>